Cleaning of gridded dat automatically works
