+++ dnora/export/exporter.py
@@ -62,13 +62,17 @@ class DataExporter:
         obj_type = data_type_from_string(obj_type)
         writer_function = self._setup_export(obj_type, writer, dry_run)
 
-        if not self._silent:
-            msg.header(
-                writer_function,
-                f"Writing {obj_type.name} data from {self.model[obj_type].name}",
-            )
-
         if not self.dry_run():
+            if self.model.get(obj_type) is None:
+                msg.info(f"No {obj_type.name} data exists. Won't export anything.")
+                return
+
+            if not self._silent:
+                msg.header(
+                    writer_function,
+                    f"Writing {obj_type.name} data from {self.model[obj_type].name}",
+                )
+
             try:  # GeneralWritingFunction might not have this method defined
                 wanted_convention = writer_function.convention()
             except AttributeError:
@@ -76,6 +80,15 @@ class DataExporter:
 
             if obj_type in [DnoraDataType.SPECTRA, DnoraDataType.SPECTRA1D]:
                 self.model[obj_type].set_convention(wanted_convention)
+        else:
+            if not self._silent:
+                if self.model.get(obj_type) is None:
+                    msg.info(f"Writing {obj_type.name} data")
+                else:
+                    msg.header(
+                        writer_function,
+                        f"Writing {obj_type.name} data from {self.model[obj_type].name}",
+                    )
 
         self._export_object(
             obj_type,
@@ -91,9 +104,6 @@ class DataExporter:
         self, obj_type: DnoraDataType, writer_function, dry_run: bool
     ) -> WriterFunction:
         self._dry_run = dry_run
-
-        if self.model[obj_type] is None:
-            return None
 
         writer_function = writer_function or self._get_writer(obj_type)
 
@@ -113,10 +123,6 @@ class DataExporter:
         **kwargs,
     ) -> list[str]:
         # Controls generation of file names using the proper defaults etc.
-        if writer_function is None:
-            msg.info(f"No {obj_type.name} data exists. Won't export anything.")
-            return
-
         format = format or self._get_default_format()
         file_object = FileNames(
             format=format,
