+++ dnora/read/generic/generic_readers.py
@@ -14,6 +14,7 @@ from dnora.type_manager.dnora_objects import dnora_objects
 from dnora.aux_funcs import get_url
 from dnora import msg, utils
 from .constant_funcs import create_constant_array, print_constant_values
+from dnora.read.data_var_decoding import read_data_vars, compile_data_vars
 from copy import copy
 from geo_skeletons import GriddedSkeleton
 
@@ -307,220 +308,7 @@ class ConstantData(SpectralDataReader):
 
         meta_dict = {"zone_number": zone_number, "zone_letter": zone_letter}
 
-        # Create metaparameters based on standard short names
-        # metaparameter_dict = gp.parameter_funcs.create_parameter_dict(data_dict.keys())
-
         return coord_dict, data_dict, meta_dict
-
-
-# class ConstantGriddedData(DataReader):
-#     _class_default_var_values = {"default": 1.0}
-#     _class_default_coord_values = {"default": np.linspace(0.0, 10.0, 11)}
-
-#     def __init__(self, vars: dict = None, coords: dict = None):
-#         """E.g. ConstantGrid(vars={'u':1, 'v':2})"""
-#         vars = vars or {}
-#         self._default_var_values = copy(self._class_default_var_values)
-#         self._default_var_values.update(vars)
-
-#         coords = coords or {}
-#         self._default_coord_values = copy(self._class_default_coord_values)
-#         self._default_coord_values.update(coords)
-
-#     def _caching_strategy(self) -> CachingStrategy:
-#         return CachingStrategy.DontCacheMe
-
-#     def __call__(
-#         self,
-#         obj_type: DnoraDataType,
-#         grid,
-#         start_time,
-#         end_time,
-#         source: DataSource,
-#         folder: str,
-#         new_vars: dict = None,
-#         time: list[str] = None,
-#         dt: int = None,
-#         **kwargs,
-#     ):
-#         """Variables that are a part of the class can be given as a keyword, e.g. u=1.
-#         To add variables dynamically, give the values as a dict new_vars, e.g. new_vars={'hs': 6}
-
-#         time: determines start time if constant values given as an array
-#         dt [h]: Can be given instad of time to determine start times of array values
-
-#         If ModelRun is defined for 2020-01-01 00:00 - 2020-01-01 23:00, set 12 + 12 hour wind speed with:
-#         u=[1, 2, v=[1,2], time=['2020-01-01 00:00', '2020-01-01 12:00']
-
-#         OR
-
-#         u=[1, 2, v=[1,2], dt=12
-#         """
-
-#         coord_dict = grid.coord_dict(strict=True)
-
-#         # Determine size of the object to create
-#         object_coords = dnora_objects.get(obj_type)._coord_manager.added_coords()
-#         obj_size = []
-
-#         # Time is first if exists
-#         if "time" in object_coords:
-#             time_vec = pd.date_range(start=start_time, end=end_time, freq="h")
-#             coord_dict["time"] = time_vec
-#             obj_size.append(len(time_vec))
-#             object_coords.remove(time)
-#         else:
-#             time_vec = None
-#         # Then spatial coords
-#         obj_size.append(grid.ny())
-#         obj_size.append(grid.nx())
-
-#         # Then all other coords
-#         for obj_coord in object_coords:
-#             default = self._default_coord_values.get("default")
-#             val = kwargs.get(
-#                 "obj_coord", self._default_coord_values.get(obj_coord, default)
-#             )
-#             coord_dict[obj_coord] = val
-#             obj_size.append(len(val))
-
-#         obj_size = tuple(obj_size)
-
-#         # Get values for data variables
-#         object_vars = dnora_objects.get(obj_type)._coord_manager.added_vars().keys()
-#         data_val_dict = create_datvar_val_dict(
-#             kwargs, default_values=self._default_var_values, object_vars=object_vars
-#         )
-#         data_dict = {}
-#         for key, val in data_val_dict.items():
-#             data_dict[key] = create_constant_array(val, time, obj_size, time_vec)
-
-#         print_constant_values(data_dict, obj_type, time_vec)
-
-#         meta_dict = {}
-
-#         # Create metaparameters based on standard short names
-#         metaparameter_dict = create_metaparameter_dict(data_dict.keys())
-
-#         return coord_dict, data_dict, meta_dict, metaparameter_dict
-
-
-# class ConstantPointData(SpectralDataReader):
-#     def __init__(
-#         self,
-#         fp: float | None = 0.3,
-#         dirp: int | None = 0,
-#         convention: SpectralConvention | str = SpectralConvention.OCEAN,
-#         **kwargs,
-#     ):
-#         """**kwargs are possible extra coordinates
-#         E.g. ConstantGrid(z=np.linspace(0,10,11))
-
-#         fp/dirp used when importing spectra only
-#         """
-#         self.extra_coords = kwargs
-#         if self.extra_coords.get("freq") is None:
-#             self.extra_coords["freq"] = np.linspace(0.1, 1, 10)
-#         if self.extra_coords.get("dirs") is None:
-#             self.extra_coords["dirs"] = np.linspace(0, 350, 36)
-
-#         self.dirp = dirp
-#         self.fp = fp
-#         self.provided_convention = convention
-
-#     def get_coordinates(self, grid, start_time, source, folder, **kwargs):
-#         lon_all, lat_all = grid.lonlat(strict=True)
-#         x_all, y_all = grid.xy(strict=True)
-
-#         return {"lon": lon_all, "lat": lat_all, "x": x_all, "y": y_all}
-
-#     def __call__(
-#         self,
-#         obj_type: DnoraDataType,
-#         grid,
-#         start_time,
-#         end_time,
-#         source: DataSource,
-#         folder: str,
-#         inds,
-#         time=None,
-#         dt=None,
-#         force_type: str = "",
-#         **kwargs,
-#     ):
-#         coord_dict = {}
-
-#         obj_size = []
-#         # Time is always first coord if exists
-#         if "time" in dnora_objects.get(obj_type)._coord_manager.added_coords():
-#             time_coord = self.extra_coords.get(
-#                 "time", pd.date_range(start=start_time, end=end_time, freq="h").values
-#             )
-#             coord_dict["time"] = time_coord
-#             obj_size.append(len(time_coord))
-#         else:
-#             time_coord = None
-#         # Inds always second (or first if time doesn't exist)
-#         obj_size.append(len(inds))
-
-#         for added_coord in dnora_objects.get(obj_type)._coord_manager.added_coords():
-#             if added_coord != "time":
-#                 coord_val = self.extra_coords.get(added_coord, np.linspace(1, 100, 100))
-#                 coord_dict[added_coord] = coord_val
-#                 obj_size.append(len(coord_val))
-
-#         obj_size = tuple(obj_size)
-
-#         all_coordinates = self.get_coordinates(grid, start_time, source, "")
-#         if all_coordinates.get("lon") is not None:
-#             coord_dict["lon"] = all_coordinates.get("lon")[inds]
-#         if all_coordinates.get("lat") is not None:
-#             coord_dict["lat"] = all_coordinates.get("lat")[inds]
-#         if all_coordinates.get("x") is not None:
-#             coord_dict["x"] = all_coordinates.get("x")[inds]
-#         if all_coordinates.get("y") is not None:
-#             coord_dict["y"] = all_coordinates.get("y")[inds]
-
-#         if self.fp is not None and obj_type in [
-#             DnoraDataType.SPECTRA,
-#             DnoraDataType.SPECTRA1D,
-#         ]:
-#             """Set everything except dominant frequency to 0"""
-#             non_fp_ind = np.argmin(
-#                 np.abs(self.extra_coords.get("freq") - self.fp)
-#             ) != np.arange(len(self.extra_coords.get("freq")))
-#         else:
-#             non_fp_ind = np.zeros(len(self.extra_coords.get("freq"))).astype(bool)
-
-#         if self.dirp is not None and obj_type == DnoraDataType.SPECTRA:
-#             """Set everything except dominant direction to 0"""
-#             non_dirp_ind = np.argmin(
-#                 np.abs(self.extra_coords.get("dirs") - self.dirp)
-#             ) != np.arange(len(self.extra_coords.get("dirs")))
-#         else:
-#             non_dirp_ind = np.zeros(len(self.extra_coords.get("dirs"))).astype(bool)
-
-#         variables = dnora_objects.get(obj_type)._coord_manager.added_vars().keys()
-
-#         data_dict = create_constant_data_dict(
-#             variables, kwargs, obj_size, start_time, end_time, time
-#         )
-#         print_constant_values(data_dict, obj_type, time_coord)
-#         for key in data_dict:
-#             if obj_type == DnoraDataType.SPECTRA:
-#                 data_dict[key][:, :, non_fp_ind, :] = 0
-#                 data_dict[key][:, :, :, non_dirp_ind] = 0
-#             elif obj_type == DnoraDataType.SPECTRA1D:
-#                 data_dict[key][:, :, non_fp_ind] = 0
-
-#         if obj_type == DnoraDataType.SPECTRA:
-#             self.set_convention(self.provided_convention)
-#         if obj_type == DnoraDataType.SPECTRA1D:
-#             self.set_convention(convert_2d_to_1d(self.provided_convention))
-
-#         meta_dict = {}
-#         metaparameter_dict = create_metaparameter_dict(data_dict.keys())
-#         return coord_dict, data_dict, meta_dict, metaparameter_dict
 
 
 def create_datvar_val_dict(kwargs, default_values, object_vars):
