+++ dnora/read/waveseries/waveseries_readers.py
@@ -26,8 +26,7 @@ from dnora.type_manager.data_sources import DataSource
 from dnora.wave_parameters.parameters import get_function
 from dnora.read.ds_read_functions import read_ds_list, read_first_ds
 from functools import partial
-
-# from dnora.type_manager.dnora_objects import dnora_objects
+from dnora.read.data_var_decoding import read_data_vars, compile_data_vars
 
 
 def ds_xarray_read(start_time, end_time, url):
@@ -232,7 +231,8 @@ class WW3Unstruct(PointDataReader):
         source: DataSource,
         folder: str,
         filename: str,
-        inds,
+        inds: list[int],
+        obj_data_vars: list[str],
         data_vars: list[str] = None,
         force_names: str = None,
         decode_cf: bool = None,
@@ -254,8 +254,9 @@ class WW3Unstruct(PointDataReader):
         if data_vars is None:
             data_vars = self._data_vars
 
-        # if not data_vars:
-        #     data_vars = dnora_objects.get(obj_type).non_coord_objects()
+        # If no data variables have been provided, read the ones that might be prsent in the class
+        if not data_vars:
+            data_vars = obj_data_vars
 
         folder = self._folder(folder, source)
         filename = self._filename(filename, source)
@@ -317,102 +318,3 @@ class WW3Unstruct(PointDataReader):
         meta_dict = ds.attrs
 
         return coord_dict, data_dict, meta_dict
-
-
-def compile_data_vars(data_vars, aliases: dict):
-    """Compiles data_vars accounting for possible aliases etc."""
-    new_vars = []
-    for var in data_vars:
-        name, param = gp.decode(var)
-        if param is not None:
-            name = param.standard_name()
-        if aliases.get(name) is not None:
-            if gp.is_gp_class(param):
-                param = param(aliases.get(name))  # Create e.g. Dirp('dp')
-            else:
-                param.name = aliases.get(name)
-            msg.plain(
-                f"Identified that {type(param).__name__} is called {aliases.get(name)} in the source!"
-            )
-            new_vars.append(param)
-        else:
-            new_vars.append(name)
-    return new_vars
-
-
-def read_data_vars(
-    data_vars: list[str, gp.metaparameter.MetaParameter],
-    ds: xr.Dataset,
-    keep_gp_names: bool,
-    keep_source_names: bool,
-    decode_cf: bool,
-) -> dict:
-    """Read the given variables from the xarray dataset
-
-    Uninitilized geo-parameter, e.g. gp.wave.Hs
-        - Decode using standard-name. If not found, try default name of parameter (e.g. 'hs').
-    Uninitilized geo-parameter, e.g. gp.wave.Hs('hsig')
-        - Decode using standard-name. If not found, try given name of parameter ('hsig').
-    String, e.g. 'hs'
-        - Try to read that variable from dataset
-
-    keep_source_names = True
-        - Keep the variable name that is used by the data source even if a geo-parameter is given
-        - E.g. gp.wave.Hs can match to 'hsig' by standard_name and 'hsig' will be used ('hs' if false)
-
-    keep_gp_names = True
-        - Create the variable names using the default class name even if it is initialized
-        - E.g. gp.wave.Hs('swh') matches 'swh' variable with no metadata, but the variable created is still called 'hs'
-
-    decode_cf = True
-        - Try to decode standard_name from source and create geo-parameters if only string given
-        - E.g. given 'hs' will try to find standard name in metadata of 'hs' and create gp.wave.Hs
-    """
-
-    if keep_gp_names and keep_source_names:
-        raise ValueError("keep_source_names and keep_gp_names cannot both be True!")
-    data_dict = {}
-    for var in data_vars:
-        var, param = gp.decode(var, init=True)
-        if param is not None:
-            var_in_ds = param.find_me_in_ds(ds)
-            if var_in_ds is None:
-                var_in_ds = var
-            if keep_gp_names:
-                param.name = gp.get(param.standard_name()).name
-            elif keep_source_names:
-                param.name = var_in_ds
-
-            if hasattr(ds[var_in_ds], "standard_name"):
-                std_name = ds[var_in_ds].standard_name
-            else:
-                std_name = ""
-
-            msg.plain(
-                f"[{type(param).__name__}('{param.name}')] << '{var_in_ds}' {std_name}"
-            )
-            data_dict[param] = ds.get(var_in_ds).data
-        elif ds.get(var) is not None:
-            if hasattr(ds[var], "standard_name"):
-                std_name = ds[var].standard_name
-            else:
-                std_name = ""
-            if std_name and decode_cf:
-                param = gp.get(std_name)
-            else:
-                param = None
-            if param is not None:
-                if keep_gp_names:
-                    param = param()
-                else:
-                    param = param(var)
-                msg.plain(
-                    f"[{type(param).__name__}('{param.name}')] << '{var}' {std_name}"
-                )
-                data_dict[param] = ds.get(var).data
-            else:
-                msg.plain(f"'{var}' << {var}")
-                data_dict[var] = ds.get(var).data
-        else:
-            msg.plain(f"'{var}' unknown. Skipping.")
-    return data_dict
