+++ environment.yml
@@ -2,7 +2,7 @@ name: dnora2
 channels:
   - conda-forge
 dependencies:
-  - geo-skeletons>=0.16.6
+  - geo-skeletons>=0.17.1
   - geo-parameters>=0.4.5
   - python=>3.9
   - numpy>=1.17
